requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
distro==1.9.0
dlib==19.24.5
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
openai==1.40.0
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rich==13.7.1
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tqdm==4.66.5
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1


# PyTorch for Linux with CUDA 11.1
torch==2.4.0+cu121; platform_system == "Linux"
torchvision==0.19.0+cu121; platform_system == "Linux"

# PyTorch for macOS (CPU only)
torch==2.4.0; platform_system == "Darwin"
torchvision==0.19.0; platform_system == "Darwin"
